refactor editorpass
