+++ Editor/GizmoPass.h
@@ -1,0 +1,38 @@
+#pragma once
+
+#include "Pass.h"
+#include "Gizmo.h"
+
+namespace ToolKit
+{
+  namespace Editor
+  {
+
+    struct GizmoPassParams
+    {
+      Viewport* Viewport = nullptr;
+      BillboardRawPtrArray GizmoArray;
+    };
+
+    class GizmoPass : public Pass
+    {
+     public:
+      GizmoPass();
+      explicit GizmoPass(const GizmoPassParams& params);
+
+      void Render() override;
+      void PreRender() override;
+      void PostRender() override;
+
+     public:
+      GizmoPassParams m_params;
+
+     private:
+      SpherePtr m_depthMaskSphere = nullptr;
+      Camera* m_camera            = nullptr;
+    };
+
+    typedef std::shared_ptr<GizmoPass> GizmoPassPtr;
+
+  } // namespace Editor
+} // namespace ToolKit
